LPS-96442 Inline BuiltInJSModule.
